# Replace striped executor with serial executor.