--- a/californium-core/src/main/java/org/eclipse/californium/core/network/CoapEndpoint.java
+++ b/californium-core/src/main/java/org/eclipse/californium/core/network/CoapEndpoint.java
@@ -58,7 +58,9 @@
  *    Achim Kraus (Bosch Software Innovations GmbH) - reject messages with MID only
  *                                                    (therefore tcp messages are not rejected)
  *    Achim Kraus (Bosch Software Innovations GmbH) - setup retransmitResponse for notifies
- *    Achim Kraus (Bosch Software Innovations GmbH) - forward onConnect
+ *    Achim Kraus (Bosch Software Innovations GmbH) - forward onConnecting and onDtlsRetransmission
+ *    Achim Kraus (Bosch Software Innovations GmbH) - replace striped executor
+ *                                                    with serial executor
  ******************************************************************************/
 package org.eclipse.californium.core.network;
 
@@ -70,6 +72,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.Executor;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.ScheduledExecutorService;
@@ -111,8 +114,6 @@ import org.eclipse.californium.elements.UDPConnector;
 import org.eclipse.californium.elements.util.DaemonThreadFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import eu.javaspecialists.tjsn.concurrency.stripedexecutor.StripedExecutorService;
 
 /**
  * Endpoint encapsulates the stack that executes the CoAP protocol. Endpoint
@@ -197,11 +198,7 @@ public class CoapEndpoint implements Endpoint {
 	private final DataParser parser;
 
 	/** The executor to run tasks for this endpoint and its layers */
-	private ScheduledExecutorService executor;
-	/**
-	 * Striped executor facade based on the {@link #executor}.
-	 */
-	private volatile Executor exchangeExecutor;
+	private ExecutorService executor;
 
 	/** Indicates if the endpoint has been started */
 	private boolean started;
@@ -405,22 +402,22 @@ public class CoapEndpoint implements Endpoint {
 			}
 		}
 
-		Executor exchangeExecutionHandler = new Executor() {
+		final Executor exchangeExecutionHandler = new Executor() {
 
 			@Override
 			public void execute(Runnable command) {
-				Executor executor = exchangeExecutor;
-				if (executor == null) {
+				final Executor exchangeExecutor = executor;
+				if (exchangeExecutor == null) {
 					LOGGER.error("Executor not ready for exchanges!",
 							new Throwable("exchange execution failed!"));
 				} else {
-					executor.execute(command);
+					exchangeExecutor.execute(command);
 				}
 			}
 		};
 
 		this.connector.setEndpointContextMatcher(endpointContextMatcher);
-		LOGGER.info("{} uses {}", new Object[] { getClass().getSimpleName(), endpointContextMatcher.getName() });
+		LOGGER.info("{} uses {}", getClass().getSimpleName(), endpointContextMatcher.getName());
 
 		this.coapstack = coapStackFactory.createCoapStack(connector.getProtocol(), config, new OutboxImpl());
 
@@ -470,7 +467,6 @@ public class CoapEndpoint implements Endpoint {
 				}
 			});
 		}
-		exchangeExecutor = new StripedExecutorService(executor);
 
 		try {
 			LOGGER.debug("Starting endpoint at {}", getUri());
@@ -548,9 +544,13 @@ public class CoapEndpoint implements Endpoint {
 
 	@Override
 	public synchronized void setExecutor(final ScheduledExecutorService executor) {
-		// TODO: don't we need to stop and shut down the previous executor?
-		this.executor = executor;
-		this.coapstack.setExecutor(executor);
+		if (this.executor != executor) {
+			if (started) {
+				throw new IllegalStateException("endpoint already started!");
+			}
+			this.executor = executor;
+			this.coapstack.setExecutor(executor);
+		}
 	}
 
 	@Override
@@ -592,11 +592,11 @@ public class CoapEndpoint implements Endpoint {
 	public void sendRequest(final Request request) {
 		// create context, if not already set
 		request.prepareDestinationContext();
-		Exchange exchange = new Exchange(request, Origin.LOCAL, exchangeExecutor);
-		exchange.execute(new StripedExchangeJob(exchange) {
+		final Exchange exchange = new Exchange(request, Origin.LOCAL, executor);
+		exchange.execute(new Runnable() {
 
 			@Override
-			public void runStriped() {
+			public void run() {
 				coapstack.sendRequest(exchange, request);
 			}
 		});
@@ -604,30 +604,39 @@ public class CoapEndpoint implements Endpoint {
 
 	@Override
 	public void sendResponse(final Exchange exchange, final Response response) {
-		exchange.execute(new StripedExchangeJob(exchange) {
-
-			@Override
-			public void runStriped() {
-				if (exchange.getRequest().getOptions().hasObserve()) {
-					// observe- or cancel-observe-requests may have multiple responses
-					// when observes are finished, the last response has no longer an
-					// observe option. Therefore check the request for it.
-					exchange.retransmitResponse();
+		if (exchange.checkOwner()) {
+			// send response while processing exchange.
+			coapstack.sendResponse(exchange, response);
+		} else {
+			exchange.execute(new Runnable() {
+				@Override
+				public void run() {
+					if (exchange.getRequest().getOptions().hasObserve()) {
+						// observe- or cancel-observe-requests may have multiple responses
+						// when observes are finished, the last response has no longer an
+						// observe option. Therefore check the request for it.
+						exchange.retransmitResponse();
+					}
+					coapstack.sendResponse(exchange, response);
 				}
-				coapstack.sendResponse(exchange, response);
-			}
-		});
+			});
+		}
 	}
 
 	@Override
 	public void sendEmptyMessage(final Exchange exchange, final EmptyMessage message) {
-		exchange.execute(new StripedExchangeJob(exchange) {
+		if (exchange.checkOwner()) {
+			// send response while processing exchange.
+			coapstack.sendEmptyMessage(exchange, message);
+		} else {
+			exchange.execute(new Runnable() {
 
-			@Override
-			public void runStriped() {
-				coapstack.sendEmptyMessage(exchange, message);
-			}
-		});
+				@Override
+				public void run() {
+					coapstack.sendEmptyMessage(exchange, message);
+				}
+			});
+		}
 	}
 
 	/**
@@ -911,12 +920,12 @@ public class CoapEndpoint implements Endpoint {
 
 			// MessageInterceptor might have canceled
 			if (!request.isCanceled()) {
-				Exchange exchange = matcher.receiveRequest(request);
+				final Exchange exchange = matcher.receiveRequest(request);
 				if (exchange != null) {
-					exchangeExecutor.execute(new StripedExchangeJob(exchange) {
+					exchange.execute(new Runnable() {
 
 						@Override
-						public void runStriped() {
+						public void run() {
 							exchange.setEndpoint(CoapEndpoint.this);
 							coapstack.receiveRequest(exchange, request);
 						}
@@ -938,13 +947,13 @@ public class CoapEndpoint implements Endpoint {
 
 			// MessageInterceptor might have canceled
 			if (!response.isCanceled()) {
-				Exchange exchange = matcher.receiveResponse(response);
+				final Exchange exchange = matcher.receiveResponse(response);
 				if (exchange != null) {
-					exchangeExecutor.execute(new StripedExchangeJob(exchange) {
+					exchange.execute(new Runnable() {
 
 						@Override
-						public void runStriped() {
-							// entered striped execution.
+						public void run() {
+							// entered serial execution.
 							// recheck, if the response still match the exchange
 							// and the exchange is not changed in the meantime
 							if (exchange.checkCurrentResponse(response)) {
@@ -954,7 +963,7 @@ public class CoapEndpoint implements Endpoint {
 							} else if (!response.isDuplicate() && response.getType() != Type.ACK) {
 								LOGGER.debug("rejecting not longer matchable response from {}",
 										response.getSourceContext());
-							//	reject(response);
+								// reject(response);
 							} else {
 								LOGGER.debug("not longer matched response {}", response);
 							}
@@ -986,13 +995,13 @@ public class CoapEndpoint implements Endpoint {
 					LOGGER.debug("responding to ping from {}", message.getSourceContext());
 					reject(message);
 				} else {
-					Exchange exchange = matcher.receiveEmptyMessage(message);
+					final Exchange exchange = matcher.receiveEmptyMessage(message);
 					if (exchange != null) {
-						exchangeExecutor.execute(new StripedExchangeJob(exchange) {
+						exchange.execute(new Runnable() {
 
 							@Override
-							public void runStriped() {
-								// entered striped execution.
+							public void run() {
+								// entered serial execution.
 								// recheck, it the empty message still match the exchange
 								// and the exchange is not changed in the meantime
 								if (exchange.checkMID(message.getMID())) {