… californium-core/src/main/java/org/eclipse/californium/core/network/StripedExchangeJob.java
… californium-core/src/main/java/org/eclipse/californium/core/network/stack/ObserveLayer.java
… californium-core/src/main/java/org/eclipse/californium/core/network/stack/ReliabilityLayer.java
… demo-apps/cf-extplugtest-server/src/main/java/org/eclipse/californium/extplugtests/ExtendedTestServer.java
… element-connector/src/main/java/eu/javaspecialists/tjsn/concurrency/stripedexecutor/StripedCallable.java
… element-connector/src/main/java/eu/javaspecialists/tjsn/concurrency/stripedexecutor/StripedExecutorService.java
… element-connector/src/main/java/eu/javaspecialists/tjsn/concurrency/stripedexecutor/StripedObject.java
… element-connector/src/main/java/eu/javaspecialists/tjsn/concurrency/stripedexecutor/StripedRunnable.java
… element-connector/src/main/java/org/eclipse/californium/elements/util/SerialExecutor.java
… scandium-core/src/main/java/org/eclipse/californium/scandium/DTLSConnector.java
… scandium-core/src/test/java/org/eclipse/californium/scandium/DTLSConnectorTest.java