@sbernard31 2018-09-05
I just finished to read this refactoring.
@boaks could you remind me the purpose of this ? Did you remarks some performances benefits ?