@boaks 2018-09-05
Together with the executor utils (PR #728) it's about 30% faster for DTLS.
Just checkout the branch and even without the other PR, it should be faster, but I didn't measure it on its own.