@sbernard31 2018-09-05
I tested it using : `coaps://localhost:5784/benchmark?rlen=40 2000 1000`

### without this PR
<details>
  <summary>2000000 requests in 136712ms, 14629 reqs/s.</summary>

```
Create 2000 secure benchmark clients, expect to send 2000000 request overall to coaps://localhost:5784/benchmark?rlen=40
Benchmark clients created.
Benchmark started.
259355 requests (25935 reqs/s, 3374 retransmissions (1,30%), 0 transmission errors (0,00%), 2000 clients)
401261 requests (14190 reqs/s, 2222 retransmissions (1,57%), 0 transmission errors (0,00%), 2000 clients)
517605 requests (11634 reqs/s, 2067 retransmissions (1,78%), 0 transmission errors (0,00%), 2000 clients)
695476 requests (17787 reqs/s, 321 retransmissions (0,18%), 0 transmission errors (0,00%), 2000 clients)
835252 requests (13977 reqs/s, 1357 retransmissions (0,97%), 0 transmission errors (0,00%), 2000 clients)
998579 requests (16332 reqs/s, 502 retransmissions (0,31%), 0 transmission errors (0,00%), 2000 clients)
1154896 requests (15631 reqs/s, 93 retransmissions (0,06%), 0 transmission errors (0,00%), 2000 clients)
1325039 requests (17014 reqs/s, 200 retransmissions (0,12%), 0 transmission errors (0,00%), 2000 clients)
1463272 requests (13823 reqs/s, 778 retransmissions (0,56%), 0 transmission errors (0,00%), 2000 clients)
1614484 requests (15121 reqs/s, 0 retransmissions (0,00%), 0 transmission errors (0,00%), 2000 clients)
1773375 requests (15889 reqs/s, 1952 retransmissions (1,23%), 0 transmission errors (0,00%), 2000 clients)
1925278 requests (15190 reqs/s, 214 retransmissions (0,14%), 0 transmission errors (0,00%), 2000 clients)
2000 benchmark clients finished.
2000000 requests sent, 2000000 expected
2000000 requests in 136712ms, 14629 reqs/s
13227 retransmissions (0,66%)
  1 clients with 754 requests.
 15 clients with 761 to 799 requests.
 77 clients with 801 to 839 requests.
194 clients with 841 to 879 requests.
268 clients with 880 to 919 requests.
315 clients with 920 to 959 requests.
279 clients with 960 to 999 requests.
212 clients with 1000 to 1039 requests.
173 clients with 1040 to 1079 requests.
151 clients with 1080 to 1119 requests.
102 clients with 1120 to 1159 requests.
 82 clients with 1160 to 1199 requests.
 41 clients with 1201 to 1237 requests.
 29 clients with 1242 to 1278 requests.
 25 clients with 1281 to 1319 requests.
 10 clients with 1321 to 1351 requests.
 16 clients with 1360 to 1395 requests.
  4 clients with 1401 to 1436 requests.
  3 clients with 1441 to 1470 requests.
  2 clients with 1490 to 1493 requests.
  1 clients with 1536 requests.
17:51:22.586 WARN [DTLSConnector]: Shutdown DTLS connector on [0.0.0.0/0.0.0.0:54896] not terminated in time!
17:51:22.583 WARN [DTLSConnector]: Shutdown DTLS connector on [0.0.0.0/0.0.0.0:45153] not terminated in time!
17:51:22.586 WARN [DTLSConnector]: Shutdown DTLS connector on [0.0.0.0/0.0.0.0:50839] not terminated in time!
17:51:22.586 WARN [DTLSConnector]: Shutdown DTLS connector on [0.0.0.0/0.0.0.0:46395] not terminated in time!
```
</details> 
<details>
  <summary>2000000 requests in 135236ms, 14788 reqs/s.</summary>

```
Create 2000 secure benchmark clients, expect to send 2000000 request overall to coaps://localhost:5784/benchmark?rlen=40
Benchmark clients created.
Benchmark started.
324418 requests (32441 reqs/s, 2938 retransmissions (0,91%), 0 transmission errors (0,00%), 2000 clients)
469152 requests (14473 reqs/s, 414 retransmissions (0,29%), 0 transmission errors (0,00%), 2000 clients)
587552 requests (11840 reqs/s, 2083 retransmissions (1,76%), 0 transmission errors (0,00%), 2000 clients)
766759 requests (17920 reqs/s, 339 retransmissions (0,19%), 0 transmission errors (0,00%), 2000 clients)
903620 requests (13686 reqs/s, 824 retransmissions (0,60%), 0 transmission errors (0,00%), 2000 clients)
1040425 requests (13680 reqs/s, 0 retransmissions (0,00%), 0 transmission errors (0,00%), 2000 clients)
1222944 requests (18251 reqs/s, 370 retransmissions (0,20%), 0 transmission errors (0,00%), 2000 clients)
1362260 requests (13931 reqs/s, 1079 retransmissions (0,77%), 0 transmission errors (0,00%), 2000 clients)
1538343 requests (17608 reqs/s, 143 retransmissions (0,08%), 0 transmission errors (0,00%), 2000 clients)
1681943 requests (14360 reqs/s, 466 retransmissions (0,32%), 0 transmission errors (0,00%), 2000 clients)
1861213 requests (17927 reqs/s, 51 retransmissions (0,03%), 0 transmission errors (0,00%), 2000 clients)
2000 benchmark clients finished.
2000000 requests sent, 2000000 expected
2000000 requests in 135236ms, 14788 reqs/s
9023 retransmissions (0,45%)
  1 clients with 770 requests.
 19 clients with 800 to 838 requests.
 98 clients with 840 to 879 requests.
285 clients with 880 to 919 requests.
425 clients with 920 to 959 requests.
429 clients with 960 to 999 requests.
257 clients with 1000 to 1039 requests.
152 clients with 1040 to 1079 requests.
 86 clients with 1080 to 1119 requests.
 56 clients with 1120 to 1159 requests.
 41 clients with 1160 to 1199 requests.
 47 clients with 1200 to 1237 requests.
 33 clients with 1241 to 1277 requests.
 39 clients with 1280 to 1318 requests.
 19 clients with 1325 to 1358 requests.
  8 clients with 1373 to 1397 requests.
  2 clients with 1487 to 1507 requests.
  2 clients with 1543 to 1558 requests.
  1 clients with 1616 requests.
17:45:01.719 WARN [DTLSConnector]: Shutdown DTLS connector on [0.0.0.0/0.0.0.0:42482] not terminated in time!
17:45:01.716 WARN [DTLSConnector]: Shutdown DTLS connector on [0.0.0.0/0.0.0.0:43289] not terminated in time!
17:45:01.716 WARN [DTLSConnector]: Shutdown DTLS connector on [0.0.0.0/0.0.0.0:47205] not terminated in time!
17:45:01.716 WARN [DTLSConnector]: Shutdown DTLS connector on [0.0.0.0/0.0.0.0:47477] not terminated in time!

```
</details>    
    
 
### with this PR :
<details>
  <summary>2000000 requests in 139285ms, 14359 reqs/s.</summary>

```
reate 2000 secure benchmark clients, expect to send 2000000 request overall to coaps://localhost:5784/benchmark?rlen=40
Benchmark clients created.
Benchmark started.
261017 requests (26101 reqs/s, 3275 retransmissions (1,25%), 0 transmission errors (0,00%), 2000 clients)
404598 requests (14358 reqs/s, 2074 retransmissions (1,44%), 0 transmission errors (0,00%), 2000 clients)
530390 requests (12579 reqs/s, 1988 retransmissions (1,58%), 0 transmission errors (0,00%), 2000 clients)
688061 requests (15767 reqs/s, 147 retransmissions (0,09%), 0 transmission errors (0,00%), 2000 clients)
853320 requests (16525 reqs/s, 933 retransmissions (0,56%), 0 transmission errors (0,00%), 2000 clients)
998123 requests (14480 reqs/s, 612 retransmissions (0,42%), 0 transmission errors (0,00%), 2000 clients)
1129659 requests (13153 reqs/s, 1012 retransmissions (0,77%), 0 transmission errors (0,00%), 2000 clients)
1281519 requests (15186 reqs/s, 342 retransmissions (0,23%), 0 transmission errors (0,00%), 2000 clients)
1425999 requests (14448 reqs/s, 59 retransmissions (0,04%), 0 transmission errors (0,00%), 2000 clients)
1575308 requests (14930 reqs/s, 610 retransmissions (0,41%), 0 transmission errors (0,00%), 2000 clients)
1727346 requests (15203 reqs/s, 756 retransmissions (0,50%), 0 transmission errors (0,00%), 2000 clients)
1869967 requests (14262 reqs/s, 898 retransmissions (0,63%), 0 transmission errors (0,00%), 2000 clients)
2000 benchmark clients finished.
17:39:39.131 ERROR [BenchmarkClient]: failed after 1428 requests!
17:39:39.136 ERROR [ReliabilityLayer]: Exception for Exchange[L2000245, complete] in MessageObserver: connector must be started before sending messages is possible
java.lang.IllegalStateException: connector must be started before sending messages is possible
	at org.eclipse.californium.scandium.DTLSConnector.send(DTLSConnector.java:1349)
	at org.eclipse.californium.core.network.CoapEndpoint$OutboxImpl.sendRequest(CoapEndpoint.java:741)
	at org.eclipse.californium.core.network.stack.BaseCoapStack$StackBottomAdapter.sendRequest(BaseCoapStack.java:187)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer.sendRequest(ReliabilityLayer.java:108)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer$1.retransmit(ReliabilityLayer.java:104)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer$RetransmissionTask.retry(ReliabilityLayer.java:448)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer$RetransmissionTask.access$0(ReliabilityLayer.java:405)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer$RetransmissionTask$1.run(ReliabilityLayer.java:400)
	at org.eclipse.californium.elements.util.SerialExecutor$1.run(SerialExecutor.java:272)
	at java.util.concurrent.Executors$RunnableAdapter.call(Executors.java:511)
	at java.util.concurrent.FutureTask.run(FutureTask.java:266)
	at java.util.concurrent.ScheduledThreadPoolExecutor$ScheduledFutureTask.access$201(ScheduledThreadPoolExecutor.java:180)
	at java.util.concurrent.ScheduledThreadPoolExecutor$ScheduledFutureTask.run(ScheduledThreadPoolExecutor.java:293)
	at java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1149)
	at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:624)
	at java.lang.Thread.run(Thread.java:748)
2000000 requests sent, 2000000 expected
2000000 requests in 139285ms, 14359 reqs/s
12763 retransmissions (0,64%)
  3 clients with 667 to 671 requests.
  4 clients with 681 to 707 requests.
  5 clients with 725 to 756 requests.
  5 clients with 766 to 796 requests.
 51 clients with 801 to 839 requests.
109 clients with 840 to 879 requests.
188 clients with 880 to 919 requests.
315 clients with 920 to 959 requests.
407 clients with 960 to 999 requests.
355 clients with 1000 to 1039 requests.
237 clients with 1040 to 1079 requests.
118 clients with 1080 to 1119 requests.
 68 clients with 1120 to 1157 requests.
 46 clients with 1160 to 1198 requests.
 39 clients with 1200 to 1237 requests.
 21 clients with 1240 to 1277 requests.
 14 clients with 1283 to 1312 requests.
  5 clients with 1320 to 1353 requests.
  4 clients with 1364 to 1379 requests.
  1 clients with 1428 requests.
  1 clients with 1456 requests.
  1 clients with 1502 requests.
  3 clients with 1769 to 1794 requests.
```
</details> 
<details>
  <summary>2000000 requests in 141808ms, 14103 reqs/s</summary>

```
Create 2000 secure benchmark clients, expect to send 2000000 request overall to coaps://localhost:5784/benchmark?rlen=40
Benchmark clients created.
Benchmark started.
275591 requests (27559 reqs/s, 3930 retransmissions (1,43%), 0 transmission errors (0,00%), 2000 clients)
401129 requests (12553 reqs/s, 2563 retransmissions (2,04%), 0 transmission errors (0,00%), 2000 clients)
518471 requests (11734 reqs/s, 1987 retransmissions (1,69%), 0 transmission errors (0,00%), 2000 clients)
680688 requests (16221 reqs/s, 666 retransmissions (0,41%), 0 transmission errors (0,00%), 2000 clients)
832431 requests (15174 reqs/s, 594 retransmissions (0,39%), 0 transmission errors (0,00%), 2000 clients)
971347 requests (13891 reqs/s, 1547 retransmissions (1,11%), 0 transmission errors (0,00%), 2000 clients)
1117028 requests (14568 reqs/s, 692 retransmissions (0,48%), 0 transmission errors (0,00%), 2000 clients)
1257717 requests (14068 reqs/s, 707 retransmissions (0,50%), 0 transmission errors (0,00%), 2000 clients)
1415564 requests (15784 reqs/s, 118 retransmissions (0,07%), 0 transmission errors (0,00%), 2000 clients)
1553278 requests (13771 reqs/s, 810 retransmissions (0,59%), 0 transmission errors (0,00%), 2000 clients)
1690631 requests (13735 reqs/s, 591 retransmissions (0,43%), 0 transmission errors (0,00%), 2000 clients)
1860410 requests (16977 reqs/s, 1334 retransmissions (0,79%), 0 transmission errors (0,00%), 2000 clients)
2000 benchmark clients finished.
17:54:51.351 ERROR [BenchmarkClient]: failed after 1371 requests!
17:54:51.355 ERROR [ReliabilityLayer]: Exception for Exchange[L2000923, complete] in MessageObserver: connector must be started before sending messages is possible
java.lang.IllegalStateException: connector must be started before sending messages is possible
	at org.eclipse.californium.scandium.DTLSConnector.send(DTLSConnector.java:1349)
	at org.eclipse.californium.core.network.CoapEndpoint$OutboxImpl.sendRequest(CoapEndpoint.java:741)
	at org.eclipse.californium.core.network.stack.BaseCoapStack$StackBottomAdapter.sendRequest(BaseCoapStack.java:187)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer.sendRequest(ReliabilityLayer.java:108)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer$1.retransmit(ReliabilityLayer.java:104)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer$RetransmissionTask.retry(ReliabilityLayer.java:448)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer$RetransmissionTask.access$0(ReliabilityLayer.java:405)
	at org.eclipse.californium.core.network.stack.ReliabilityLayer$RetransmissionTask$1.run(ReliabilityLayer.java:400)
	at org.eclipse.californium.elements.util.SerialExecutor$1.run(SerialExecutor.java:272)
	at java.util.concurrent.Executors$RunnableAdapter.call(Executors.java:511)
	at java.util.concurrent.FutureTask.run(FutureTask.java:266)
	at java.util.concurrent.ScheduledThreadPoolExecutor$ScheduledFutureTask.access$201(ScheduledThreadPoolExecutor.java:180)
	at java.util.concurrent.ScheduledThreadPoolExecutor$ScheduledFutureTask.run(ScheduledThreadPoolExecutor.java:293)
	at java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1149)
	at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:624)
	at java.lang.Thread.run(Thread.java:748)
2000000 requests sent, 2000000 expected
2000000 requests in 141808ms, 14103 reqs/s
16045 retransmissions (0,80%)
  1 clients with 635 requests.
  1 clients with 676 requests.
  7 clients with 685 to 714 requests.
  2 clients with 733 to 740 requests.
  6 clients with 763 to 799 requests.
 36 clients with 806 to 839 requests.
172 clients with 840 to 879 requests.
321 clients with 880 to 919 requests.
331 clients with 920 to 959 requests.
241 clients with 960 to 999 requests.
206 clients with 1000 to 1039 requests.
200 clients with 1040 to 1079 requests.
179 clients with 1080 to 1119 requests.
108 clients with 1120 to 1159 requests.
 77 clients with 1160 to 1198 requests.
 40 clients with 1200 to 1239 requests.
 26 clients with 1240 to 1279 requests.
 22 clients with 1280 to 1318 requests.
 19 clients with 1320 to 1356 requests.
  3 clients with 1364 to 1388 requests.
  1 clients with 1406 requests.
  1 clients with 1454 requests.
```
</details>  
   

There is an illegalStateException with this PR and lot of warning (not terminated in time!) at this end of the tests.