@boaks 2018-09-05
My results:

coaps://localhost:5784/benchmark?rlen=40 500 1000

Without:
500000 requests in 19673ms, 25415 reqs/s

With:
500000 requests in 17327ms, 28856 reqs/s
117 retransmissions (0,02%)

coap://localhost:5784/benchmark?rlen=40 500 1000

Without:
1000000 requests in 18501ms, 54051 reqs/s
1648 retransmissions (0,16%)

With:
1000000 requests in 15691ms, 63730 reqs/s
1467 retransmissions (0,15%)

On my system, I let the server run and start the benchmark twice, the second run is mostly faster. It's also important to have either enough heap or an other deduplicator (I use -d64 -Xmx8g -XX:+UseG1GC).

I can't start more client without the executor optimization of the other PR.
My "final" version uses the optimized executor and multiple receiver threads.
But thought you want the smaller PRs, I isolated this replacement.
The changes are mostly effect the start/stop therefore it's hard to present them in parallel and therefore their either already merged together or a few parallel.