@boaks 2018-09-06
With the last benchmark from yesterday I got aware, that testing on linux uses only 1 sender and 1 receiver thread. Using the branch "all_executor_optimizations" and removing all Californium?.properties to setup the new defaults results in #processor sender threads and #processors/2 receiver threads for the benchmark server. I repeated the first benchmark also with that multi-thread setup:

without this PR:
coap:
1000000 requests in 20827ms, 48014 reqs/s
1661 retransmissions (0,17%)
coaps:
500000 requests in 15392ms, 32484 reqs/s
1364 retransmissions (0,27%)

with this PR:
coap:
1000000 requests in 16084ms, 62173 reqs/s
1371 retransmissions (0,14%)
coaps
500000 requests in 10914ms, 45812 reqs/s
670 retransmissions (0,13%)

coap: 29%
coaps: 41%