@sbernard31 2018-09-06
I rebench (coaps://localhost:5784/benchmark?rlen=40 500 1000) a bit using you advice.

(removing californium.properties and running twice and using -d64 -Xmx8g -XX:+UseG1GC)

Without this PR : 
500000 requests in 32589ms, 15342 reqs/s
1305 retransmissions (0,26%)
500000 requests in 31668ms, 15788 reqs/s
753 retransmissions (0,15%)

With this PR : 
500000 requests in 37202ms, 13440 reqs/s
1713 retransmissions (0,34%)
500000 requests in 36361ms, 13750 reqs/s
1443 retransmissions (0,29%)

After the bench I also give a try with visualvm to look at memory/cpu consumption and it seems a bit higher with this PR. (especially memory which go up to ~2500MB againt ~2000MB) I'm aware this is not a precise metrics, so I just share this as a hint.